feat: add get seller by id service
