+++ src/api/services/seller/index.js
@@ -1,11 +1,13 @@
 const CreateSellerService = require('./create-seller.service');
 const UpdateSellerService = require('./update-seller.service');
 const GetAllSellerService = require('./get-all-seller.service');
-const GetSellerByTagService = require('./get-by-tag.service');
+const GetSellerByTagService = require('./get-by-tag-id.service');
+const GetSellerByIdService = require('./get-by-id.service');
 
 module.exports = {
     CreateSellerService,
     UpdateSellerService,
     GetAllSellerService,
     GetSellerByTagService,
+    GetSellerByIdService,
 };
